anime list filter with id instead login'

+++ src/helpers/deleteAnimeFromList.js
@@ -4,12 +4,12 @@ const deleteAnimeFromList = async(req, res, model) => {
         if(!user) {
             return res.status(401).json({message:"JWT ERROR"})
         }
-        const isAnimeInList = await model.findOne({animeId: req.params.animeId})
+        const isAnimeInList = await model.findOne({userId: user._id, animeId: req.params.animeId})
         if(!isAnimeInList) {
-            return res.status(404).json(`Anime with id '${req.params.animeId}' not found in user '${user.login}' lists`)
+            return res.status(404).json(`Anime with id '${req.params.animeId}' not found in user '${user.id}' lists`)
         }
-        await model.findOneAndDelete({animeId: req.params.animeId})
-        const newFavoriteList = await model.find({userLogin: user.login})
+        await model.findOneAndDelete({userId: user.id, animeId: req.params.animeId})
+        const newFavoriteList = await model.find({userId: user.id})
         // const newFavoriteList = await model.find()
         // newFavoriteList.save()
         return res.json(newFavoriteList)
